Memory card manager: should be functionally complete. Feedback will be appreciated. Major new: insert-from/remove-to file system files. New: context menu and buttons for the new functionality (including copy which was only available via drag and drop). Mod: slight text changes (finally should be ready for translations). Mod: slight cleanup of functionality: - 'Copy' changed to 'Duplicate', and can be invoked on any card, or via drag and drop to an empty port. - drag and drop of cards: from a port to the filesystem = eject card (not swap). - drag and drop of cards: from the filesystem to a port = insert the card to the port.
While it's functionally complete (I hope), I'm not 100% happy with the GUI. Better GUI would have been 2 separate lists (ps2 ports, unused cards) and cards can be moved between the lists. However, creating another list control, laying it out and handling the interactions between the two lists would have been a headache (even more than it already was). The major case which is hampered by the combined list is when the user has many cards on the filesystem, and he wants to insert it to a port, but the ports are way up the list so a scroll is required. This case is handled by the "Insert card" button which is available for any card on the filesystem (prompts the user to select a target port).

Let me know what you think.